Update to ember-cli 0.2.6 and minor tweaks
