+++ addon/initializers/visible.js
