+++ app/initializers/visible.js
@@ -1,0 +1,1 @@
+export { default, initialize } from 'ember-cli-visible/initializers/visible';
