+++ tests/dummy/app/router.js
@@ -5,5 +5,7 @@ var Router = Ember.Router.extend({
   location: config.locationType
 });
 
-export default Router.map(function() {
+Router.map(function() {
 });
+
+export default Router;
